Fixed linear classifier loss and identified tokenizer issue
The linear model loss was switched to cross entropy, dtype of y was updated for the new loss, and loss.backward was updated for multiple passes.

Also identified an issue with the tokenizer. Max_len is applied to our embeddings, but not to the pos or srl tags. We need to find a way to determine which words correspond to the first max_len=75 tokens in order to only keep their associated pos/srl tags in our target variable.